Add basic client implementation
The client hasn't finished establishing connection yet. It should send a
request to the server, which I think shouldn't go through the public
Write() interface, and then wait for the server to return its connection
ID, which I have temporarily made it go through the public Read()
interface.

The reason that Read() is public is the main thread has a distinct
handler for each type of message read, in which a MsgConnect can quickly
be identified and returned to user (the connection establishment
routine). However, if the handler of Write() happens before actually
Write(), so we build our own MsgConnect (with id = 0) and then use UDP's
interface to write.